Use folder name to define release version instead of including in json
There's no real use in having it in the json since the folder already
gives the version number. It just opens up the possibility to have
conflicting versions defined between the folder name and the json file.

I will take advantage of this later to make querying a little easier to
filter by not having to read/decode the info.json files to determine the
version number.